requirements.txt
# Deep Learning frameworks
tensorflow
torch==2.6.0
keras
transformers==4.50.3

# Data manipulation and analysis
numpy
pandas==2.2.3
scikit-learn==1.6.1
scipy==1.15.2
db-dtypes
plotly

# NLP tools and datasets
huggingface-hub==0.30.1
datasets==3.5.0
evaluate==0.4.3
tokenizers==0.21.1
nltk==3.9.1
newspaper3k==0.2.8
wordcloud==1.9.4

# API and Servers
fastapi
uvicorn
streamlit
google-cloud-bigquery

# Visualization
matplotlib==3.10.1
seaborn==0.13.2
pillow==11.1.0

# Utilities
tqdm==4.67.1
joblib==1.4.2
pydantic==2.11.1
requests==2.32.3
fsspec==2024.12.0
PyYAML==6.0.2
loguru==0.7.3

# Secondary dependencies 
filelock==3.18.0
accelerate==1.6.0
safetensors==0.5.3